swarm support

+++ Ring/Ring/Features/ContactRequests/ContactRequestsViewModel.swift
@@ -37,6 +37,7 @@ class ContactRequestsViewModel: Stateable, ViewModel {
     let nameService: NameService
     let presenceService: PresenceService
     let profileService: ProfilesService
+    let requestsService: RequestsService
 
     private let disposeBag = DisposeBag()
     private let log = SwiftyBeaver.self
@@ -50,22 +51,23 @@ class ContactRequestsViewModel: Stateable, ViewModel {
         self.nameService = injectionBag.nameService
         self.presenceService = injectionBag.presenceService
         self.profileService = injectionBag.profileService
+        self.requestsService = injectionBag.requestsService
 
         self.injectionBag = injectionBag
     }
 
-    lazy var contactRequestItemsNotFiltered: Observable<[ContactRequestItem]> = {
-        return self.contactsService.contactRequests
+    lazy var contactRequestItemsNotFiltered: Observable<[RequestItem]> = {
+        return self.requestsService.requests
             .asObservable()
-            .map({ [weak self] contactRequests in
+            .map({ [weak self] requests in
                 guard let self = self else { return [] }
-                return contactRequests
+                return requests
                     .filter { $0.accountId == self.accountsService.currentAccount?.id }
                     .sorted { $0.receivedDate > $1.receivedDate }
                     .map { contactRequest in
-                        let item = ContactRequestItem(withContactRequest: contactRequest,
-                                                      profileService: self.profileService,
-                                                      contactService: self.contactsService)
+                        let item = RequestItem(withRequest: contactRequest,
+                                               profileService: self.profileService,
+                                               contactService: self.contactsService)
                         self.lookupUserName(withItem: item)
                         return item
                     }
@@ -74,9 +76,9 @@ class ContactRequestsViewModel: Stateable, ViewModel {
 
     let filter = BehaviorRelay(value: "")
 
-    lazy var contactRequestItems: Observable<[ContactRequestItem]> = {
+    lazy var contactRequestItems: Observable<[RequestItem]> = {
         return Observable
-            .combineLatest(contactRequestItemsNotFiltered, filter.asObservable()) { (requests, filter) -> [ContactRequestItem] in
+            .combineLatest(contactRequestItemsNotFiltered, filter.asObservable()) { (requests, filter) -> [RequestItem] in
                 return requests.filter { request in
                     return request.userName.value.contains(filter) || request.profileName.value.contains(filter) || filter.isEmpty
                 }
@@ -84,7 +86,7 @@ class ContactRequestsViewModel: Stateable, ViewModel {
     }()
 
     lazy var hasInvitations: Observable<Bool> = {
-        return self.contactsService.contactRequests
+        return self.requestsService.requests
             .asObservable()
             .map({ [weak self] contactRequests in
                 return contactRequests
@@ -95,41 +97,51 @@ class ContactRequestsViewModel: Stateable, ViewModel {
             })
     }()
 
-    func accept(withItem item: ContactRequestItem) -> Observable<Void> {
-        let acceptCompleted = self.contactsService.accept(contactRequest: item.contactRequest, withAccount: self.accountsService.currentAccount!)
-        self.presenceService.subscribeBuddy(withAccountId: (self.accountsService.currentAccount?.id)!,
-                                            withUri: item.contactRequest.ringId,
-                                            withFlag: true)
-        return acceptCompleted.asObservable()
+    func accept(withItem item: RequestItem) -> Observable<Void> {
+        if item.request.type == .contact || (item.request.participants.count == 1 && self.contactsService.contact(withUri: item.request.participants.first!.jamiId) == nil) {
+            let acceptCompleted = self.requestsService.acceptContactRequest(jamiId: item.request.participants.first!.jamiId, withAccount: item.request.accountId)
+            self.presenceService.subscribeBuddy(withAccountId: item.request.accountId,
+                                                withUri: item.request.participants.first!.jamiId,
+                                                withFlag: true)
+            return acceptCompleted.asObservable()
+        }
+        return self.requestsService.acceptConverversationRequest(conversationId: item.request.conversationId, withAccount: item.request.accountId)
     }
 
-    func discard(withItem item: ContactRequestItem) -> Observable<Void> {
-        return self.contactsService.discard(from: item.contactRequest.ringId,
-                                            withAccountId: item.contactRequest.accountId)
+    func discard(withItem item: RequestItem) -> Observable<Void> {
+        // for conversation we discard contact request if it one to one conversation and contact not added yet
+        if item.request.type == .contact || (item.request.participants.count == 1 && self.contactsService.contact(withUri: item.request.participants.first!.jamiId) == nil) {
+            return self.requestsService.discardContactRequest(jamiId: item.request.participants.first!.jamiId, withAccount: item.request.accountId)
+        }
+        return self.requestsService.discardConverversationRequest(conversationId: item.request.conversationId, withAccount: item.request.accountId)
     }
 
-    func ban(withItem item: ContactRequestItem) -> Observable<Void> {
-        let discardCompleted = self.contactsService.discard(from: item.contactRequest.ringId,
-                                                            withAccountId: item.contactRequest.accountId)
-        guard let uri = JamiURI.init(schema: URIType.ring,
-                                     infoHach: item.contactRequest.ringId)
-            .uriString else {
+    func ban(withItem item: RequestItem) -> Observable<Void> {
+        if item.request.type == .contact || (item.request.participants.count == 1 && self.contactsService.contact(withUri: item.request.participants.first!.jamiId) == nil) {
+            let discardCompleted = self.requestsService.discardContactRequest(jamiId: item.request.participants.first!.jamiId, withAccount: item.request.accountId)
+            guard let uri = JamiURI.init(schema: URIType.ring,
+                                         infoHach: item.request.participants.first!.jamiId)
+                    .uriString else {
                 return discardCompleted
-        }
-        let removeCompleted = self.contactsService.removeContact(withUri: uri,
-                                                                 ban: true,
-                                                                 withAccountId: item.contactRequest.accountId)
+            }
+            let removeCompleted = self.contactsService.removeContact(withUri: uri,
+                                                                     ban: true,
+                                                                     withAccountId: item.request.accountId)
 
-        return Observable<Void>.zip(discardCompleted, removeCompleted) { _, _ in
-            return
+            return Observable<Void>.zip(discardCompleted, removeCompleted) { _, _ in
+                return
+            }
         }
+        return self.requestsService.discardConverversationRequest(conversationId: item.request.conversationId, withAccount: item.request.accountId)
     }
 
-    private func lookupUserName(withItem item: ContactRequestItem) {
+    private func lookupUserName(withItem item: RequestItem) {
+        if !item.request.isCoredialog() { return }
+        guard let jamiId = item.request.participants.first?.jamiId else { return }
 
         self.nameService.usernameLookupStatus.asObservable()
             .filter({ lookupNameResponse in
-                return lookupNameResponse.address == item.contactRequest.ringId
+                return lookupNameResponse.address == jamiId
             })
             .subscribe(onNext: { lookupNameResponse in
                 if lookupNameResponse.state == .found && !lookupNameResponse.name.isEmpty {
@@ -143,23 +155,16 @@ class ContactRequestsViewModel: Stateable, ViewModel {
 
         self.nameService.lookupAddress(withAccount: currentAccount.id,
                                        nameserver: "",
-                                       address: item.contactRequest.ringId)
+                                       address: jamiId)
     }
 
-    func showConversation (forRingId ringId: String) {
+    func showConversation (forItem item: RequestItem) {
         let conversationViewModel = ConversationViewModel(with: self.injectionBag)
-        guard let account = accountsService.currentAccount else {
-            return
-        }
-
-        guard let uri = JamiURI(schema: URIType.ring, infoHach: ringId).uriString else {
-            return
-        }
-
-        guard let conversation = self.conversationService.findConversation(withUri: uri, withAccountId: account.id) else {
-            return
-        }
+        let conversation = ConversationModel(withId: item.request.conversationId, accountId: item.request.accountId)
+        conversationViewModel.displayName.accept(item.userName.value)
+        conversationViewModel.profileImageData.accept(item.profileImageData.value)
         conversationViewModel.conversation = BehaviorRelay<ConversationModel>(value: conversation)
+        conversationViewModel.request = item.request
         self.stateSubject.onNext(ConversationState.conversationDetail(conversationViewModel: conversationViewModel))
     }
 }
